修复CircleDeque和CircleBufferMemoryStream的索引转换问题，优化Counter类运算符返回类型

+++ include/base/container/CircleDeque.h
@@ -98,7 +98,7 @@ namespace base
 				return Size;
 			}
 
-			return _end - _begin;
+			return static_cast<uint64_t>(_end - _begin);
 		}
 
 		///
@@ -133,7 +133,7 @@ namespace base
 				throw std::runtime_error{CODE_POS_STR + "队列为空，无法退队。"};
 			}
 
-			int64_t index = _end - 1;
+			int64_t index = static_cast<uint64_t>(_end - 1);
 			T ret{std::move(Buffer()[index])};
 			Buffer()[index].~T();
 			_end--;
@@ -154,7 +154,7 @@ namespace base
 				return false;
 			}
 
-			int64_t index = _end - 1;
+			int64_t index = static_cast<uint64_t>(_end - 1);
 			out = std::move(Buffer()[index]);
 			Buffer()[index].~T();
 			_end--;
